xiaohongshu-collector-skill/requirements.txt
# 小红书笔记收集器 - Python依赖

# 核心依赖
pyyaml>=6.0
python-docx>=1.1.0

# 开发依赖（可选）
# black>=24.0.0
# flake8>=7.0.0
# pytest>=8.0.0